Make the following changes:
,

Find this element:

  <div class="container mx-auto px-4">

with the text content:

    Our FacilitiesModern Patient RoomsComfortable and well-equipped private and semi-private rooms.Advanced LaboratoryState-of-the-art diagnostic testing facilities.Intensive Care Unit (ICU)Specialized care for critically ill patients with continuous monitoring.Outpatient ClinicsConvenient access to specialist consultations and follow-ups.Emergency Department24/7 emergency care services with experienced staff.

Relevant files:

And change it as follows:

all the cards in our facilities section should open when clicked to show a modal that has the picture and more details about that facility. the needed changes should be applied to the admin panel to accommodate the new feature.