Make the following changes:
,

Find this element:

  <div class="container mx-auto px-4">

with the text content:

    GalleryPhotos (6)Videos (2)

Relevant files:

And change it as follows:

when I click on the pictures, it should show all the pictures that are in the gallery with a sliding show meaning the gallery should be functional on click and the playback for the videos also should be functional on click. on hover the videos in the gallary should be previewed.