Make the following changes:
,

Find this element:

  <div class="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-8">

with the text content:

    Free Health Check-up CampJuly 20, 2024Join us for a free health screening event next Saturday. Limited slots available.Read MoreGrace Hospital Opens New Cardiology WingJuly 15, 2024Our expanded cardiology department offers cutting-edge treatments and diagnostics.Read MoreDr. Emily Carter Joins Grace HospitalJuly 10, 2024We are pleased to welcome renowned neurologist Dr. Carter to our expert team.Read More

Relevant files:

And change it as follows:

when clicked on "read more" of the news and events section it should display the blog with more description. there should also be a progress bar that shows how much of the news the person have read.