Make the following changes:
,

Find this element:

  <Footer>

with the text content:

    Grace Hospital© 2025 Grace Hospital. All rights reserved.Privacy Policy|Terms of Service|Contact Us

Relevant files:

    - https://6000-idx-studio-1746511554605.cluster-3gc7bglotjgwuxlqpiut7yyqt4.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

the footer should also include a social media icons with link to the social media of the hospital, the social medias should be telegram, tiktok and facebook. also create an input feild for these social media links in the site setting section of the admin panel.