Make the following changes:
,

Find this element:

  <FacilitiesSection>

with the text content:

    Our FacilitiesModern Patient RoomsComfortable and well-equipped for recovery.Advanced LaboratoryState-of-the-art diagnostic testing.Intensive Care Unit (ICU)Specialized critical care 24/7.Outpatient ClinicsConvenient specialist consultations.Emergency Department24/7 immediate medical attention.

Relevant files:

    - https://6000-idx-studio-1746511554605.cluster-3gc7bglotjgwuxlqpiut7yyqt4.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

the cards in the facilities section should change the color from white to red left to right on hover. and also the icons color should change from red to blue on hover because it won't be visible when the card turns red.