Make the following changes:
,

Find this element:

  <CardContent>

with the text content:

    NameEmailMessageSend Message

Relevant files:

    - https://6000-idx-studio-1746511554605.cluster-3gc7bglotjgwuxlqpiut7yyqt4.cloudworkstations.dev/_next/static/chunks/src_e16348d8._.js
    - https://6000-idx-studio-1746511554605.cluster-3gc7bglotjgwuxlqpiut7yyqt4.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

when the customer send us a message here I want that message to be delivered to "sirawbizugelangelan@gmail.com" this email address. configure it such that all the messages from the users on the contact us section will be sent to the email I sent you.